Sort out install/copy rules for the .prf and .java files
NOTE: We must NOT install the .java files to

    $$[QT_INSTALL_PREFIX]/src/android/java/src

because the whole tree is copied into the android-build dir
and unconditionally compiled during application packaging.
Thus, we use $$[QT_INSTALL_PREFIX]/src/android/java/qml as
an installation location so that we can conditionally copy
the desired .java files to the application build directory.

Change-Id: I7d154a1df8fc9307d6588fc84c5f967c6dc42e06
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>